requirements.txt
# Install Necessary Packages
fastapi==0.110.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.27
pymysql==1.1.0
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
jinja2==3.1.2
google-generativeai
openai==2.20.0
passlib
pydantic[email]
bcrypt===3.2.0